pyproject.toml
[tool.poetry]
name = "habit tracker"
version = "0.1.0"
description = ""
authors = ["Annabse27 <annabse27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.5.0"
celery = "^5.4.0"
redis = "^5.1.1"
python-decouple = "^3.8"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
django-celery-beat = "^2.7.0"
coverage = "^7.6.3"
sentry-sdk = "^2.17.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
